% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuscluster.R
\name{consensuscluster}
\alias{consensuscluster}
\title{Title}
\usage{
consensuscluster(
  data = NULL,
  K = 2,
  B = 100,
  pItem = 0.8,
  pFeature = 1,
  clMethod = "hclust",
  dist = "euclidean",
  hclustMethod = "average",
  finalclmethod = "hclust",
  finalhclustMethod = "average",
  sparseKmeansPenalty = NULL,
  maxIterKM = 1000,
  nstart = 20
)
}
\arguments{
\item{nstart}{}
}
\description{
Title
}
