% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{criterion}
\title{Title}
\usage{
criterion(
  CM,
  clLabels,
  K,
  Silhouette = Silhouette,
  widestGap = widestGap,
  dunns = dunns,
  dunn2s = dunn2s
)
}
\arguments{
\item{CM}{A list of consensus matrix.}

\item{clLabels}{A matrix consists of clustering result corresponding to each consensus matrix.}

\item{K}{A vector denotes number of clusters for each consensus matrix.}

\item{Silhouette}{Boolean. If TRUE, compute Silhouette to choose
best number of clusters. Default is TRUE.}

\item{widestGap}{Boolean. If TRUE, compute also widest gap index to choose
best number of clusters. Default is FALSE.}

\item{dunns}{Boolean. If TRUE, compute also Dunn's index to choose best
number of clusters. Default is FALSE.}

\item{dunn2s}{Boolean. If TRUE, compute also alternative Dunn's index to
choose best number of clusters. Default is FALSE.}
}
\value{
best final K and criterion used
}
\description{
Title
}
