% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Title}
\usage{
data_generation(clustersize, info_features, total_features)
}
\arguments{
\item{clustersize}{A number or a vector denotes the number of samples for each cluster.}

\item{info_features}{Number of informative features.}

\item{total_features}{Number of total features.}
}
\description{
Title
}
