% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CM.R
\name{plot_CM}
\alias{plot_CM}
\title{Consensus clustering}
\usage{
plot_CM(data = NULL, result = NULL)
}
\arguments{
\item{data}{N X P data matrix}

\item{result}{Results derived from consensus clustering.
It consists of a list of several consensus matrix and corresponding clustering label results.}
}
\value{
heatmap for consensus matrix from the result or heatmap for dataset
}
\description{
This function allows to perform consensus clustering using the k-means
clustering algorithm, for a fixed number of clusters. We consider the number
of clusters K to be fixed.
}
\examples{
library(intCC)
load(system.file("extdata", "exampleData.RData", package = "intCC"))
normData=exampleData[[1]]
normRes=consensuscluster(normData,K=3,B=1000,pItem = 0.8,pFeature = 0.8,
clMethod ="kmeans",finalclmethod="pam")
plot_CM(result=normRes)
plot_CM(data=normData,result=normRes)
}
