% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightCal.R
\name{weightcal}
\alias{weightcal}
\title{Title}
\usage{
weightcal(result)
}
\arguments{
\item{result}{Results derived from consensus clustering.
It consists of a list of several consensus matrix and corresponding clustering label results.}
}
\value{
weights for each consensus matrix
}
\description{
Title
}
\examples{
library(weightedCC)
simulation=data_generation(20,5,30)
normData=simulation[[1]]
normRes=consensuscluster(normData,K=3,B=1000,pItem = 0.8,pFeature = 0.8 ,
clMethod ="kmeans",finalclmethod="pam")
normRes1=consensuscluster(normData,K=3,B=1000,pItem = 0.8,pFeature = 0.8 ,
clMethod ="hclust",finalclmethod="pam")
res.norm <- c(list(normRes),list(normRes1))
weight.norm<- weightcal(res.norm)
}
