% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedConsensusCluster.R
\name{WeightedConsensusCluster}
\alias{WeightedConsensusCluster}
\title{Title}
\usage{
WeightedConsensusCluster(
  data,
  method = NULL,
  individualK = NULL,
  individualMaxK = 10,
  globalK = NULL,
  B = 1000,
  pItem = 0.8,
  pFeature = 1,
  ccClMethods = "k-means",
  ccDistHCs = "euclidean",
  hclustMethod = "average",
  finalclmethod = "hclust",
  finalhclustMethod = "average",
  Silhouette = TRUE,
  widestGap = FALSE,
  dunns = FALSE,
  dunn2s = FALSE
)
}
\arguments{
\item{dunn2s}{}
}
\description{
Title
}
