% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{criterion}
\title{Title}
\usage{
criterion(
  CM,
  clLabels,
  K,
  Silhouette = Silhouette,
  widestGap = widestGap,
  dunns = dunns,
  dunn2s = dunn2s
)
}
\arguments{
\item{dunn2s}{}
}
\description{
Title
}
