% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightCal.R
\name{weightcal}
\alias{weightcal}
\title{Title}
\usage{
weightcal(result)
}
\arguments{
\item{result}{Results derived from consensus clustering.
It consists of a list of several consensus matrix and corresponding clustering label results.}
}
\value{
weights for each consensus matrix
}
\description{
Title
}
